refactor(forge): split Forgejo merge-request methods into forgejo_mr.go (#1025)
Drop the free-standing section header '--- Merge requests, issue timeline,
notes, token introspection (M4) ---': it named concerns now split across
forgejo_mr.go, forgejo_labels.go, forgejo_issues.go and forgejo_auth.go, so it
documented no single remaining block (M2 precedent).
